+++ src/tiledb/cloud/bioimg/__init__.py
@@ -4,13 +4,14 @@ from .helpers import get_embeddings_uris
 from .helpers import serialize_filter
 from .helpers import batch
 from .helpers import scale_calc
-
-# from .types import SupportedExtensions, EMBEDDINGS
+from .types import SupportedExtensions, EMBEDDINGS
 
 __all__ = ("ingest", 
            "export",
            "get_embeddings_uris",
            "serialize_filter",
-           "batch"
-           "scale_calc"
+           "batch",
+           "scale_calc",
+           SupportedExtensions,
+           EMBEDDINGS
            )
