+++ src/tiledb/cloud/bioimg/types.py
@@ -1,0 +1,9 @@
+import enum
+class EMBEDDINGS(enum.Enum):
+    RESNET = enum.auto()
+
+class SupportedExtensions(enum.Enum):
+    TIFF: str = ".tiff"
+    TIF: str = ".tif"
+    SVS: str = ".svs"
+    TDB: str = ".tdb"
